+++ src/main/java/com/dragonminez/common/config/RaceCharacterConfig.java
@@ -18,7 +18,7 @@ public class RaceCharacterConfig {
     private String customModel = "";
 
 	@SerializedName("defaultModelScaling")
-	private double defaultModelScaling = 0.9375;
+	private float defaultModelScaling = 0.9375f;
 
     @SerializedName("defaultBodyType")
     private int defaultBodyType = 0;
@@ -62,7 +62,7 @@ public class RaceCharacterConfig {
     public boolean hasGender() { return hasGender; }
     public boolean useVanillaSkin() { return useVanillaSkin; }
 	public String getCustomModel() { return customModel; }
-	public double getDefaultModelScaling() { return defaultModelScaling; }
+	public float getDefaultModelScaling() { return defaultModelScaling; }
     public int getDefaultBodyType() { return defaultBodyType; }
     public int getDefaultHairType() { return defaultHairType; }
     public int getDefaultEyesType() { return defaultEyesType; }
@@ -80,7 +80,7 @@ public class RaceCharacterConfig {
     public void setHasGender(boolean hasGender) { this.hasGender = hasGender; }
     public void setUseVanillaSkin(boolean useVanillaSkin) { this.useVanillaSkin = useVanillaSkin; }
 	public void setCustomModel(String customModel) { this.customModel = customModel; }
-	public void setDefaultModelScaling(double defaultModelScaling) { this.defaultModelScaling = defaultModelScaling; }
+	public void setDefaultModelScaling(float defaultModelScaling) { this.defaultModelScaling = defaultModelScaling; }
     public void setDefaultBodyType(int defaultBodyType) { this.defaultBodyType = defaultBodyType; }
     public void setDefaultHairType(int defaultHairType) { this.defaultHairType = defaultHairType; }
     public void setDefaultEyesType(int defaultEyesType) { this.defaultEyesType = defaultEyesType; }
